Backend/requirements.txt
fastapi
uvicorn
pydantic[dotenv]
firebase-admin
sqlalchemy
sqlite-utils
pandas
numpy
scikit-learn
tensorflow
python-dotenv
twilio
requests
httpx
pandas
numpy
scikit-learn
joblib
xgboost            
apscheduler
requests
python-dateutil

python-multipart